nat: update due to change in NAT mappings used count JSON output
The NAT code has been updated to track the address/port mappings
per-protocol (TCP/UDP/Other), and the JSON was updated to show
individual values. The MIB value of natAddrMapAddrUsed, from
.1.3.6.1.2.1.123.1.4.1.19.X.X, takes the value of "used_ts".  The code
has been updated to work with the JSON output before the change, and
after the change so that the package does not need branched.  As the MIB
does not distinguish the protocol, with the new JSON output, the code has
been updated to use the sum of "used_ts" values (across the protocols)
for natAddrMapAddrUsed.

Dependencies have been updated to depend on vyatta-dataplane-op-npf-op-0
or vyatta-dataplane-op-npf-op-1, as it can work with either version.

VRVDR-50718