MIBMisc: add GETNEXT fast lookup table
Currently a GETNEXT operation searches the sorted tree from the first
object. Therefore a GETNEXT operation takes progressively longer as
we walk the tree since eg. to retrieve the last object we need to
walk the entire tree. For a large MIB (eg. tens of thousands of
objects) this can be crippling.

To address this, build a lookup table in sort_mib_keys(). This table
maps an OID to the next OID in the tree. In get_next_oid() we lookup
this table with the input OID to obtain the next OID which we then
use to lookup the data tree - completing the operation in constant
time. If no entry was found in the lookup table then we iterate the
sorted tree as before. This fallback to iteration should only occur
in limited situations however, for example the first GETNEXT of a
MIB walk, and in this case we want the first object anyway.

VRVDR-51003