Assume MIB trees are built roughly sorted
sort_mib_keys() currently uses the keys of the $mib_tree entry hash
as the input to generate the sorted list of OIDs, used for GETNEXT
operations.

The trouble with this approach is that hash keys are returned in a
non-deterministic order. For large MIBs this can result in the sort
operation being costly.

If we instead assume that the subagent is inserting data into the
tree in, at least roughly, sorted order we can speed up the final
sort by using the OIDs in insertion order as the sort input.

For a MIB with 22,000 objects this change reduces the sort time from
4 seconds to < 1 second.

VRVDR-51003