Restart reason log is noisy
The reboot reason service is starting way
before the system has been configured, and
checking every 2 seconds if snmpd is running
before sending the trap. SNMP may not be
configured on the system. Have the service
start after system-configure service, and if
snmpd is running, wait for some time before
sending the trap for network to be reachable.

VRVDR-55779